+++ public/layers/TestLayer.js
@@ -1,6 +1,6 @@
 import {Component, Layer} from "src/gui";
 import {Vector2} from "src/math";
-import Button from "../components/ButtonSC.js";
+import Button from "../components/Button.js";
 
 export default class TestLayer extends Layer {
 	/** @override */
@@ -11,7 +11,6 @@ export default class TestLayer extends Layer {
 				margin: new Vector2(0, 0),
 				uv: new Vector2(0, 0),
 				width: 200,
-				disabled: false,
 			}),
 		];
 	}
