+++ public/main.js
@@ -1,9 +1,9 @@
 import Instance from "src/instance";
 import TextureGenerator from "src/generator";
 import {GUI, GUIRenderer} from "src/gui";
-import MainMenuLayer from "./layers/MainMenuLayer.js";
-// import TestLayer from "./layers/TestLayer.js";
-import Button from "./components/Button.js";
+// import MainMenuLayer from "./layers/MainMenuLayer.js";
+import TestLayer from "./layers/TestLayer.js";
+import Button from "./components/_Button.js";
 
 /** @todo Fix undefined instance on throw */
 
@@ -38,7 +38,7 @@ try {
 		gui.renderer.loadButtonTextures(buttonWidths);
 	}
 
-	gui.push(new MainMenuLayer());
+	gui.push(new TestLayer());
 
 	instance.startLoop();
 } catch (error) {
