+++ src/gui/components/VisualComponent.js
@@ -1,5 +1,5 @@
 import {Component} from "../index.js";
-import {Matrix3, Vector2} from "../../math/index.js";
+import {Vector2} from "../../math/index.js";
 import Texture from "../../Texture.js";
 import {extend} from "../../utils/index.js";
 
@@ -7,11 +7,8 @@ const TEXTURE_SIZE = new Vector2(256, 256);
 
 /**
  * @extends Component
- * @param {{
- *    uv: Vector2
- * }}
  */
-export function VisualComponent({uv}) {
+export function VisualComponent() {
 	Component.apply(this, arguments);
 
 	/** @type {Subcomponent[]} */
@@ -36,16 +33,16 @@ export function VisualComponent({uv}) {
 	/** @param {Texture} value */
 	this.setTexture = value => void (texture = value);
 
-	/** @returns {Matrix3} */
-	this.getTextureMatrix = () => Matrix3
-		.translate(this.getUV().divide(TEXTURE_SIZE))
-		.scale(this.getSize().divide(TEXTURE_SIZE));
+	// /** @returns {Matrix3} */
+	// this.getTextureMatrix = () => Matrix3
+	// .translate(this.getUV().divide(TEXTURE_SIZE))
+	// .scale(this.getSize().divide(TEXTURE_SIZE));
 
-	/** @returns {Vector2} */
-	this.getUV = () => uv;
+	// /** @returns {Vector2} */
+	// this.getUV = () => uv;
 
-	/** @param {Vector2} value */
-	this.setUV = value => void (uv = value);
+	// /** @param {Vector2} value */
+	// this.setUV = value => void (uv = value);
 }
 
 extend(VisualComponent, Component);
